requirements.txt
openai==0.27.4
gradio==3.24.1
langchain==0.0.131
python-dotenv